Updates to the Settings Window
One can now setup both the list view and the alert view columns in a GUI fashion in their tabs in the Settings Window.

One can now set a fixed manual position in the configuration and the Settings Window. When set, the manual position will override any incoming NMEA position messages.

It is now possible to specify a log time in the Settings Window.

Some small fixes in the default configuration.