Initial implementation of an identification DB
IDDB is a simple way to use data from previous runs of the program. For objects that only have position reports, it will try to fill in the name, imo number and the callsign from a database in memory.

This changes the memory db by adding two more tables: iddb and metadata. Only the table iddb is used for now.

When the program starts, it tries to load a DB file containing id data from a previous run. Whenever an update of the column imo takes place in the data table, it will also incorporate the changes to the iddb table. When displaying data in the GUI, the program will look up data from both the data and the iddb table and "merge" the information. It can also save the changes to the iddb table to a file.