Reimplementation of list views + other bug fixes
This changes the list views to be virtual list controls. This means that only items visible on the screen are redrawn at an update, which uses much less CPU than the old implementation.

Sorting of list views are now done by the wxpython ColumnSorterMixin. This means that we do not need to query the SQL-db, which means (at least on Linux) that sorting is much faster now. There are a visual indicator for the sorted column in the form of an ugly arrow.

Added a remark column that is to be used for displaying remarks on each object loaded from a file. The file loading mechanism for this is yet to be written.

Bugfix: The Detail Window would not display any data on some objects.

Bugfix: The network server would get objects in its queue until the whole system ran out of memory if no client was connected. This has now been fixed so that the queue only can contain 500 objects at maximum. This bug was discovered only after I managed to crash a harddrive (swapping heavily for several days). Beware! This software can actually kill your hardware. :-)