Settings can be saved and networkclient does raw
The ability to save settings has been implemented. There will be no reload when
applying or saving changes, instead the user is asked to restart the program
both when saving and applying settings.

Remade the network client to use raw data instead of homegrown format. This
means that it is now possible to connect to a raw data server anywhere (like on
the 'net).

Fixed some bugs relating to the statswindow where parserate would write to
wrong textctrl.