Fixed settings bug and input stats
This fixes a bug where the program would save settings as a boolean value where
the program expected an integer (i.e. 0 or 1). The program will now accept
boolean values from the config file.

Made exception logging to file append to file instead of overwriting it.

Fixed small issues in SerialThread. Made the stats variable an instance
variable instead of a class one... Added internal portname (i.e. A or B) to the
source tag. Added a variable that tracks if a message contains several lines
and is parsed. The stats variable will now add for each parsed line to
"parsed".

Made the according changes to NetworkClientThread.