Major update, including a remade network server.
Remade a lot of things in the NetworkServerThread. There is now a "feeder"
function that feeds each thread sending the data to clients. There was major
problems with the old way of doing things: the connected clients would share
the outgoing data!

When reading data from the remark file, try ASCII/UTF-8 first. If codec fails,
use cp1252 instead. Made degrees-symbol into a unicode object. The remark file
and the CSV file from the list export is now saved in cp1252 (a hack dedicated
to our Windows friends). Other Unicode fixes.

Fixed some code for the editing an object in the AdvancedAlertWindow.

Small fixes when inserting a new object in the SetAlertsWindow.
Other small fixes there regarding selected items.

typkod.lst renamed to typecode.lst

Added a version variable. It is displayed in the About window.

Values are now rounded in the Horizon box in the StatsWindow.

Removed some old code that were out-commented.