Continued work on backend, network client remade
Lots of backend related work, the main parts of the new
infrastructure is starting to emerge.

The network client thread can now have several connections at the
same time in an async way. The configuration options has been
updated to reflect this.

We now use "pydblite" instead of sqlite for in-memory database
usage. It is much more pythonic and is clearly faster for the many
insert and update operations in the main thread.

The raw data window has been updated to display both data,
message type number, mmsi number and source.

The fields bit and tamper has been removed from the memory db and
config options, because they are only relevant to SAAB messages,
and they are also not especially interesting.

The distance calculations now use DD format.

The own position is now in DD format, config options reflect this.

Lots of smaller things changed too.