Made logging to DB and IDDB work again
Logging to disk and loading and saving to IDDB is now working
again. There has been some changes to how the saved data is
represented, all decimal values are saved as floats, integers
as integers, strings as strings and Nonetype for empty or N/A
fields. This should save space and make searching more efficient.

Message 4, Base Station Report, is now inserted into the memory
DB and marked with the new field base_station set to True.
Such objects will not be logged to disk by a special check.

The networkclient is now non-blocking so that if we loose the
connection to one server the thread will just continue anyway.

Made deleting and grey-out of objects work again.