Made Detail Window work with new structure
The Detail Window now works with the use of standard update and
a query messages instead of the old scheme of polling. Also
removed old fields BIT and tamper in the window.

Some of the classes are now only instanced once.

Removed the feature to always ensure that a marked object is
visible in the list views.