Added support for Class B transponders and several more messages
Added support for Class B transponders and the decoding of
messages 18, 19 and 24 for these transponders.

Added message 6, 8, 12 and 14. Added a binary parser for decoding
binary payload of messages of 6 and 8. Implemented the IMO
Metrological and Hydrological message and the fundamental text
message. Adapted functions for lat/long to work with fewer bits.

The messages 6, 8, 12 and 14 is ignored by the front-end for now.

Bugfix for removal of messages in list views.

The typecode file has been updated with better descriptions.