Initial checkin if new Statistics Window + fixes
Made the Statistics Window open again, limited usability for the
time being.

We now get statistics by calling a return function in
CommHubThread.

The Raw Data Window now gets data from a special queue.

Remade the GUI grey set to a dict, where keys are MMSI numbers
and values are the distance to the object. This was needed to
be able to calculate a radio horizon.

Bug fix: we eere sending 'old' messages for all old objects,
now we just send it for "new" 'old' objects.