Made Statistics Window show input stats again
The Statistics Window will now dynamically show stats from all
inputs. The parsed rate calculations have been redone to show
the rate based on the fith last update to the window. This is
so that we avoid to get a zero rate from time to time.

Also some small changes in decode.py were made to get the
statistics working properly.