Added initial map functionality
Added a map using the FloatCanvas library in WxPython. It
displays ships and basestations on a map based on reading
shoreline data from a MapGen Format file. This adds NumPy as an
import dependency for the program.

Added configuration options for map.

Added inital integration to the program, one can open the Detail
Window by clicking or pressing a button in the map.

The implementation is based upon the FloatCanvas map demo.
Thanks a lot for such a useful library!