Added 'Zoom to object' feature
One can now choose 'Zoom to object on map' from the Detail
window or from a new right-click popup menu in the list views.
This will make the object selected, centre it on the map and
zoom to a bounding box of one degree latitude and longitude
around the object.

An error message will now be displayed if we fail to read map
data.

The Detail window is now a real frame which means that it won't
be on top of the main window.