Try to prevent unclean shutdown (on Windows)
In certain situations (on Windows) the process will still be running
when user has exited the GUI. Now we try to destroy more windows when
exiting to make sure that the GUI is exiting. A timeout check was
added so that we try to terminate the program when 30 seconds have
passed since the exit was issued, even if there are threads still
running.